pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "talentscope-engine"
version = "0.1.0"
description = "Phrase-aware CV–Job matching and candidate ranking engine with salary-aware ordering."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "pdfminer.six>=20221105",
  "python-docx>=1.1.0",
]

[project.scripts]
talentscope = "talentscope.cli:main"
